process-and-memory: Initial commit
This project requires that we right add our own syscall to the
linux kernel. The syscall needs to be named get_pid_info and
populate a struct provided by the user with information on a
process given by the second argument which is the pid of this
process. Basic setup is done with simple makefile to compile
the code together with the rest of the kernel and add function
declaration in /include/linux/syscalls.h and entry in the syscall
table.

Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>